Auto-fixes for control-card-routing (#2867)
